Rename Linear/CircularFind to just Find
It's obvious which is which based on the viewer whose subdirectory they're a part of
