+++ src/SeqViz/Linear/SeqBlock/SeqBlock.jsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import Annotations from "./Annotations/Annotations.jsx";
 import CutSiteRow from "./CutSites/CutSites.jsx";
 import IndexRow from "./Index/Index.jsx";
-import LinearFind from "./LinearFind/LinearFind.jsx";
+import LinearFind from "./Find/Find.jsx";
 import Primers from "./Primers/Primers.jsx";
 import Selection from "./Selection/Selection.jsx";
 import TranslationRows from "./Translations/Translations.jsx";
